리팩토링 및 코드 정리했다.

+++ app/src/main/java/woogear/kwon/currencyapi/model/CurrencyData.kt
@@ -3,6 +3,6 @@ package woogear.kwon.currencyapi.model
 data class CurrencyData(val success: Boolean,
                         val timeStamp: Long,
                         val source: String,
-                        val quotes: HashMap<String, Float>) {
+                        val quotes: HashMap<String, Double>) {
 
 }
